src/requirements.txt
pymongo==4.13.2
streamlit==1.46.1
openai==1.93.0
anthropic==0.57.1
python-dotenv==1.0.0
PyYAML==6.0.2
google-generativeai==0.8.5
groq==0.30.0